evaluate/requirements.txt
openai
anthropic
google-genai
python-dotenv
tqdm
pandas
